Cosmetic/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.2
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.1
django-import-export==3.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.3
tablib==3.4.0
urllib3==1.26.15
xlrd==2.0.1
xlwt==1.3.0
